NumPy update

+++ benchmarks/Python++/staticCython/matrix_multiplication.pyx
@@ -1,5 +1,5 @@
 cpdef void matmul(list A, list B, list R):
-    cdef int i, j, k
+    cdef long i, j, k
     for i in range(len(A)):
         for j in range(len(B[0])):
             for k in range(len(B)):
